requirements.txt
Django==3.2.12
django-cors-headers==3.10.1
djangorestframework==3.13.1
django-heroku==0.3.1
dj-database-url==0.5.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
whitenoise==5.3.0
djangorestframework-simplejwt==4.4.0